refactor: 아지트 관련 컨트롤러 및 서비스 파일 통합
- 아지트 생성, 수정, 삭제, 목록 조회 컨트롤러를 azit.controller.ts로 통합
- 아지트 생성, 수정, 삭제, 목록 조회 서비스를 azit.service.ts로 통합
