--- src/modules/azit/services/azit-delete.service.ts
@@ -1,69 +1,0 @@
-// src/modules/azit/services/azit-delete.service.ts
-import { injectable, inject } from "tsyringe";
-import { AzitRepository } from "../repositories/azit.repository";
-import { AzitUserRepository } from "../repositories/azit-user.repository";
-import { Result, ok, notFound, forbidden, internalServerError } from "../../../common/types/result.type";
-import { deleteFileFromS3 } from "../../../common/utils/file-util";
-import { AzitUserRole } from "@prisma/client";
-
-@injectable()
-export class AzitDeleteService {
-  constructor(
-    @inject(AzitRepository) private azitRepository: AzitRepository,
-    @inject(AzitUserRepository) private azitUserRepository: AzitUserRepository
-  ) {}
-
-  /**
-   * S3에서 파일을 삭제합니다. 실패해도 에러를 throw하지 않습니다.
-   * @param fileUrl - 삭제할 파일의 S3 URL
-   */
-  private async deleteFile(fileUrl: string | null): Promise<void> {
-    if (!fileUrl) {
-      return;
-    }
-
-    try {
-      await deleteFileFromS3(fileUrl);
-    } catch (error) {
-      // 삭제 실패해도 계속 진행
-      console.error('Failed to delete file from S3:', error);
-    }
-  }
-
-  async deleteAzit(userId: bigint, azitId: bigint): Promise<Result<void>> {
-    // 1. 아지트 존재 확인
-    const azit = await this.azitRepository.findAzitById(azitId);
-    if (!azit) {
-      return notFound({
-        message: "아지트를 찾을 수 없습니다.",
-        errorCode: "AZIT_NOT_FOUND",
-      });
-    }
-
-    // 2. 사용자가 해당 아지트의 멤버장인지 확인
-    const userRole = await this.azitUserRepository.findAzitUserRoleByUserIdAndAzitId(userId, azitId);
-    if (!userRole || userRole !== AzitUserRole.HOST) {
-      return forbidden({
-        message: "아지트 삭제는 멤버장만 가능합니다.",
-        errorCode: "AZIT_DELETE_FORBIDDEN",
-      });
-    }
-
-    // 3. S3 아이콘 파일 삭제
-    await this.deleteFile(azit.imageUrl);
-
-    // 4. 아지트 삭제
-    try {
-      await this.azitRepository.deleteAzit(azitId);
-    } catch (error) {
-      return internalServerError({
-        message: "아지트 삭제에 실패했습니다.",
-        errorCode: "AZIT_DELETE_FAILED",
-      });
-    }
-
-    return ok(undefined);
-  }
-}
-
-
